+++ tests/test_dates/test_dates_misc.py
@@ -1,5 +1,6 @@
-from datetime import datetime
+from datetime import datetime, timedelta
 
+import numpy as np
 import pandas as pd
 import pytest
 import pytz
@@ -13,7 +14,16 @@ def pytest_generate_tests(metafunc):
     if "date_func" in metafunc.fixturenames:
         metafunc.parametrize(
             "date_func",
-            ["pandas", "pydatetime", "numpy", "pandas_tz", "pydatetime_tz"],
+            [
+                "pandas",
+                "pydatetime",
+                "numpy",
+                "pandas_tz",
+                "pydatetime_tz",
+                "pandas_timedelta",
+                "pytimedelta",
+                "numpy_timedelta",
+            ],
             indirect=True,
         )
 
@@ -37,6 +47,12 @@ def date_func(request):
                 ts, pytz.timezone("Australia/Sydney")
             ).to_pydatetime()
         )
+    elif request.param == "pandas_timedelta":
+        return lambda ts: ts - pd.Timestamp(2019, 12, 31)
+    elif request.param == "pytimedelta":
+        return lambda ts: (ts - pd.Timestamp(2019, 12, 31)).to_pytimedelta()
+    elif request.param == "numpy_timedelta":
+        return lambda ts: (ts - pd.Timestamp(2019, 12, 31)).to_timedelta64()
     else:
         assert False, "should not happen"
 
@@ -50,9 +66,12 @@ def assert_expected_type(stairs, date_func):
     if stairs._data is None:
         return
     example_type = timestamp(2020, 1, 1, date_func=date_func)
-    example_type = pd.Timestamp(
-        example_type
-    )  # pandas natively converts datetimes to timestamps
+    try:  # TODO this is a hack
+        example_type = pd.Timedelta(example_type)
+    except:  # noqa
+        example_type = pd.Timestamp(
+            example_type
+        )  # pandas natively converts datetimes to timestamps
     assert all(
         [type(example_type) == type(x) for x in stairs._data.index]
     ), "Unexpected type in step points"
@@ -212,7 +231,11 @@ def s4_fix():
 
 
 def test_plot(date_func):
-    s1(date_func).plot()
+    temp = timestamp(
+        2020, 1, 1, date_func=date_func
+    )  # hack to avoid plotting with timedeltas
+    if not isinstance(temp, (pd.Timedelta, timedelta, np.timedelta64)):
+        s1(date_func).plot()
 
 
 def test_step_changes_dates(date_func):